% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_base_r_type.R
\name{as_base_r_type}
\alias{as_base_r_type}
\title{Convert xlr types to their base R type}
\usage{
as_base_r_type(x)
}
\arguments{
\item{x}{a xlr type}
}
\value{
the base type of the vector
}
\description{
\code{as_base_r_type} converts xlr types, \link{xlr_numeric}, \link{xlr_integer}, \link{xlr_percent}, and
\link{xlr_format} to their base R type. It a wrapper around \code{\link[vctrs:vec_data]{vctrs::vec_data()}} but
with a usage that is more familiar to R users.
}
