% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_integer.R
\name{beta_integer}
\alias{beta_integer}
\alias{is_beta_integer}
\alias{as_beta_integer}
\title{\code{beta_integer} vector}
\usage{
beta_integer(x = integer(), style = beta_format())

is_beta_integer(x)

as_beta_integer(x, style = beta_format())
}
\arguments{
\item{x}{A numeric vector
\itemize{
\item For \code{beta_integer()}: A numeric vector
\item For \code{is_beta_integer()}: An object to test
\item For \code{as_beta_integer()} : a vector
}}

\item{style}{Additional styling options for the vector. See \link{beta_format} for more details.}
}
\value{
An S3 vector of class \code{beta_integer}
}
\description{
This creates a double vector that represent doubles nicely, so that it is
both printed nicely, and can easily be outputted to excel in the correct format
}
\details{
Internally, \code{beta_integer} uses \code{vec_cast} to convert numeric types
to integers. Anything that \code{vec_cast} can handle so can \code{beta_integer}. Read
more about casting at \link[vctrs]{vec_cast}.
}
\examples{
library(BETAexcel)
# Create a variable to represent an integer
x <- beta_integer(2)
# This will print nicely
x
# You can change the styling, which affects how it looks when we save it as an
# excel document
x <- beta_integer(x, style = beta_format(font_size = 9, font_colour = "red"))
x
# We can also define a vector of integers
y <- beta_integer(c(1,2,3))
y
# You can convert existing data to a integer using dplyr verbs
# It formats large numbers nicely
df <- data.frame(col_1 = c(1:100*100))
df |>
  dplyr::mutate(col_pct = as_beta_integer(col_1))
# You can use as_beta_integer to convert a string in a integer
df <- data.frame(col_str = c("12","13","14"))
# now we can convert the string to a integer(), internally it uses the same
# logic as as.integer()
df |>
  dplyr::mutate(col_percent = as_beta_integer(col_str))
}
\seealso{
\code{\link[=beta_vector]{beta_vector()}}, \code{\link[=beta_percent]{beta_percent()}}, \code{\link[=beta_double]{beta_double()}}
}
