% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_vector.R
\name{beta_vector}
\alias{beta_vector}
\alias{is_beta_vector}
\alias{as_beta_vector}
\title{\code{beta_vector} vector}
\usage{
beta_vector(x = vector(), excel_format = "GENERAL", style = beta_format())

is_beta_vector(x)

as_beta_vector(x, excel_format = "GENERAL", style = beta_format())
}
\arguments{
\item{x}{A vector
\itemize{
\item For \code{beta_vector()}: A vector
\item For \code{is_beta_vector()}: An object to test
\item For \code{as_beta_vector()} : a vector
}}

\item{excel_format}{a character, the excel cell format, not validated. See
\link[openxlsx]{createStyle} argument numFmt for more details on what
you can specify.}

\item{style}{Additional styling options for the vector. See \link{beta_format} for
more details.}
}
\value{
An S3 vector of class \code{beta_vector}
}
\description{
A general container for including additional styling options with a vector.
It adds extra  formatting attribute so that it is
can easily be outputted to excel in the correct format.
}
\details{
While you can use it with \code{integer}, and \code{double} types and specifying the
associated Excel format, we recommend using \link{beta_integer}, \link{beta_double},
or \link{beta_percent} types instead.
}
\examples{
library(BETAexcel)
# Create a beta_vector object, this is used so we can add styling to an existing
# vector so that it prints nicely in excel
#
# Note currently the style will not change the style in the console
x <- beta_vector(1:100,
                 excel_format = "00.0##",
                 style = beta_format(font_size = 8))

# You can also use it so that dates are nicely printed in excel
dates <- c("02/27/92", "02/27/92", "01/14/92", "02/28/92", "02/01/92")
dates <- as.Date(dates, "\%m/\%d/\%y")
x <- beta_vector(dates,
                 # Print it as a long date in excel
                 excel_format = "LONGDATE")
# You can convert existing data to a beta_vectors using dplyr verbs
iris |>
  dplyr::mutate(iris_format = as_beta_vector(Species,
                                             "TEXT",
                                             beta_format(text_style = "italic")))
}
\seealso{
\code{\link[=beta_percent]{beta_percent()}}, \code{\link[=beta_integer]{beta_integer()}}, \code{\link[=beta_double]{beta_double()}}
}
