% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_table.R
\name{betaexcel_and_dplyr}
\alias{betaexcel_and_dplyr}
\title{BETAExcel and dplyr}
\description{
\code{beta_table()} is designed to work with dplyr verbs by default. This is so you
\code{mutate}, \code{summarise}, \code{arrange} etc. your data without losing your \code{beta_table}
information. Particularly if you have used \code{build_table} first on your data,
which outputs data as a \code{beta_table}.

The list of currently supported dplyrs verbs are: \code{arrange}, \code{distinct}, \code{filter},
\code{mutate}, \code{relocate}, \code{rename}, \code{rename_with}, \code{rowwise}, \code{select}, \code{slice},
\code{slice_head}, \code{slice_max}, \code{slice_min}, \code{slice_sample}, \code{slice_tail}, \code{summarise}.
}
