% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_to_workbook.R
\name{column_to_style}
\alias{column_to_style}
\title{convert a column into a format}
\usage{
column_to_style(col)
}
\arguments{
\item{col}{is a column from a data.frame}
}
\description{
convert a column into a format
}
