% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_from_xl.R
\name{.check_assigned_input}
\alias{.check_assigned_input}
\title{Pop-up used in \code{\link[=run_paste_from_xl]{run_paste_from_xl()}}}
\usage{
.check_assigned_input(input_name)
}
\arguments{
\item{input_name}{variable name as a string}
}
\value{
user assigned name as a string
}
\description{
If user specified a syntactically invalid variable name, provides a
suggested name and asks user to accept or change it.
}
\examples{
\dontrun{
  .check_assigned_input(' 1 bad name')
}

}
