% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_tabnames.R
\name{.check_forbidden_pad}
\alias{.check_forbidden_pad}
\title{Error for prohibited padding or more than 1 char}
\usage{
.check_forbidden_pad(pad = "")
}
\arguments{
\item{pad}{char: if pasting characters to a tab name, what character do you
want to use to pad so numbers align, (e.g. '0' for '001' or '.' for '..1'
if 3 digits of differentiation are necessary). Must be at least 1 character
but no longer. Can be a space ' ', but cannot be any of the forbidden characters
described in \code{details}}
}
\value{
a single character; throws an error if input is any of these characters \/:?*'[]
}
\description{
Error for prohibited padding or more than 1 char
}
\examples{
xlr:::.check_forbidden_pad('.')
xlr:::.check_forbidden_pad('-')
\dontrun{

# expect errors for the below: \/:?*'[]

xlr:::.check_forbidden_pad(':')
xlr:::.check_forbidden_pad('')
xlr:::.check_forbidden_pad(NA)
xlr:::.check_forbidden_pad(NULL)
xlr:::.check_forbidden_pad("\\\\")

}

}
\seealso{
\code{\link[=scrub_tabnames]{scrub_tabnames()}}
}
