% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_tabnames.R
\name{.check_tabwidth}
\alias{.check_tabwidth}
\title{Update permissible number of tabname characters}
\usage{
.check_tabwidth(width = 31, min_width = 0, quiet = FALSE)
}
\arguments{
\item{width}{the number of characters user requested for the tab/sheet names}

\item{min_width}{the minimum number of characters to permit in a tab/sheet name}

\item{quiet}{quiet overrides to the tab/sheet names}
}
\value{
between 0 and 31 where returned number is as close to user request
as possible after adjusting for reasonableness on input data
}
\description{
Adjust the specified tab/sheet names character count to be within reasonable
values and notify user with a warning when changes are made

Note: min checks should already have been done in scrub_tabnames() before passing width info
}
\examples{
.check_tabwidth(31)
.check_tabwidth(3, 4)
.check_tabwidth(34)

}
