% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_tabnames.R
\name{.fix_forbidden_tabnames}
\alias{.fix_forbidden_tabnames}
\title{Corrects tabnames by replacing reserved chars}
\usage{
.fix_forbidden_tabnames(tabnames, quiet = FALSE)
}
\arguments{
\item{tabnames}{vector of tab/sheet names}

\item{quiet}{TRUE or FALSE to suppress notices of forbidden-char replacement}
}
\value{
the input vector with forbidden characters replaced/removed
}
\description{
Replacing special characters one-for-one and notifies the user of replaced characters.
}
\note{
NA always returned as NA
NULL acts like c(), i.e c(NULL, NULL)
}
\examples{
c("' '?HisTory,*?*!@#$\%^&*()_+ '[{1'}]'[:/\\\\]''", "", NA, NULL) |>  xlr:::.fix_forbidden_tabnames()
" ''''''" |> xlr:::.fix_forbidden_tabnames()
"'''1' '" |> xlr:::.fix_forbidden_tabnames()
"'[ ]hi];'" |> xlr:::.fix_forbidden_tabnames()
"!@#$\%^&*()_+ {HI} ^" |> xlr:::.fix_forbidden_tabnames()
"'histor'y'" |> xlr:::.fix_forbidden_tabnames()
"[:/] " |> xlr:::.fix_forbidden_tabnames()
"?/a\\\\" |> xlr:::.fix_forbidden_tabnames()
"?,*?*?'" |> xlr:::.fix_forbidden_tabnames()
"[history coursework]" |> xlr:::.fix_forbidden_tabnames()
"'HISTORIC' History '''" |> xlr:::.fix_forbidden_tabnames()
"*[history' buff]*' " |> xlr:::.fix_forbidden_tabnames()
NA |> xlr:::.fix_forbidden_tabnames()
NULL |> xlr:::.fix_forbidden_tabnames()
c(NULL, NA, '[hi?]') |> xlr:::.fix_forbidden_tabnames()
c(NULL, NULL) |> xlr:::.fix_forbidden_tabnames()
enlist(NULL, NA, '[hi?]') |>  purrr::map(xlr:::.fix_forbidden_tabnames)

}
\seealso{
\code{\link[=scrub_tabnames]{scrub_tabnames()}}

\code{\link[=.forbidden_chars_replace]{.forbidden_chars_replace()}}
}
