% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_tabnames.R
\name{.forbidden_chars_replace}
\alias{.forbidden_chars_replace}
\title{Replace forbidden characters in sheet/tab or file names with a warning}
\usage{
.forbidden_chars_replace(
  names,
  pattern = "",
  replacement = "",
  pattern_text = pattern,
  repl_text = replacement,
  quiet = FALSE
)
}
\arguments{
\item{names}{vector of tab/sheet names}

\item{pattern}{a regular expression describing forbidden characters}

\item{replacement}{the text taking the place of the forbidden characters}

\item{pattern_text}{description of characters being replaced in the pattern regex, default is a repeat of pattern}

\item{repl_text}{description of the replacements for forbidden chars, default is a repeat of replacement}

\item{quiet}{TRUE or FALSE to quiet notices of forbidden char replacement}
}
\value{
the input vector with forbidden characters replaced/removed
}
\description{
Replace forbidden characters in sheet/tab or file names with a warning
}
\examples{
tab_names <- c(NA, "''''''", "[]hi];", "'HI'", "'hist''", "[:/]", "?/a\\\\\\\\", "'?,*?*'", "'[history coursework]'", "'HISTORIC'AL''''")
tab_names <- c("c(list(1-5), list(5-1))|1", "c(list(1-5), list(5-1))|2", "letters")
tab_names |> .forbidden_chars_replace(pattern = '[\\\\\\\\/:]', replacement = '-', pattern_text = '\\\\, /, or :')
tab_names |> .forbidden_chars_replace(pattern = '[?*]', replacement = '#', pattern_text = '? or *')
tab_names |> .forbidden_chars_replace(pattern = '(?i)(hist)ory', replacement = '\\\\1', pattern_text = 'history', repl_text = 'hist')
tab_names |> .forbidden_chars_replace(pattern = '\\\\[', replacement = '{', pattern_text = '[brackets]', repl_text = '{curly braces}')
tab_names |> .forbidden_chars_replace(pattern = ']', replacement = '}', quiet = TRUE)
tab_names |> .forbidden_chars_replace(pattern = '^'+', replacement = ' ', pattern_text = 'single quotes '' at tabname start or end', repl_text = 'the empty string ''')
tab_names |> .forbidden_chars_replace(pattern = ''+$', replacement = '', quiet = TRUE)
}
\seealso{
\code{\link[=.fix_forbidden_tabnames]{.fix_forbidden_tabnames()}}
}
