% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscript.R
\name{.isolate_names}
\alias{.isolate_names}
\title{Title}
\usage{
.isolate_names(.expr)
}
\arguments{
\item{.expr}{}
}
\description{
Title
}
\examples{
enlist(head(iris), head(mtcars)) |> isolate_names()
enlist(candy, !!candy, !!!candy, rlang::list2(!!!candy), list(list(`<list>` = list('something`something')))) |> encodeString() |> clipr::write_clip()
dplyr::starwars |> head() |> encodeString() |> clipr::write_clip()
enlist(head(iris), head(mtcars), head(dplyr::starwars)) |> encodeString() |> clipr::write_clip()
enlist(head(iris), head(mtcars), head(dplyr::starwars)) |> isolate_names()

}
