% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_tabnames.R
\name{.paste_names}
\alias{.paste_names}
\title{Paste name components together in order specified by paste_side}
\usage{
.paste_names(prefix_suffix, tabnames, sep = ".", paste_side = "right")
}
\arguments{
\item{prefix_suffix}{vector of padded numbers}

\item{tabnames}{vector of tab/sheet names}

\item{sep}{separator characters}

\item{paste_side}{'right' or 'left', the side to attach the unique numbering to}
}
\value{
a vector of glued tabnames
}
\description{
Paste name components together in order specified by paste_side
}
\examples{
xlr:::.paste_names('--1', c('hello','goodbye'))
xlr:::.paste_names('..1', c('hello','goodbye'), '->', 'left')
xlr:::.paste_names(NA, c('hello','goodbye'), '->', 'left')
}
