% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlist.R
\name{.remove_doublequotes}
\alias{.remove_doublequotes}
\title{Removes external set of quotes for named list if the list names have double-quotes}
\usage{
.remove_doublequotes(named_list = NULL)
}
\arguments{
\item{named_list}{list with names}
}
\value{
named list where names have no double-quotes
}
\description{
Removes external set of quotes for named list if the list names have double-quotes
}
\examples{
tibble::tibble(!!!letters) |> as.list() |> .remove_doublequotes()
}
