% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entibble.R
\name{.to_tibble}
\alias{.to_tibble}
\title{Internal function used in entibble.
Process dataframes, matrices, vectors, factors to named tibble.}
\usage{
.to_tibble(.in_data = NULL, .rowname = "rowname", .vec_name = "")
}
\arguments{
\item{.in_data}{user data to tibble}

\item{.rowname}{user specified rowname; ignored if no rownames}

\item{.vec_name}{The name of the expression; passed to \code{\link[=.vector_2_tibble]{.vector_2_tibble()}}}
}
\value{
a tibble
}
\description{
Internal function used in entibble.
Process dataframes, matrices, vectors, factors to named tibble.
}
\examples{
.to_tibble(head(mtcars), 'vehicle')
}
