% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entibble.R
\name{.vector_2_tibble}
\alias{.vector_2_tibble}
\title{Transform a named vector, factor, or list to a 1 or 2 column tibble}
\usage{
.vector_2_tibble(.in_vec, .rowname = "rowname", .vec_name = "")
}
\arguments{
\item{.in_vec}{input data}

\item{.rowname}{user specified name for rows}

\item{.vec_name}{name for vector, typically an expression}
}
\value{
tibble
}
\description{
Function called in \code{\link[=.to_tibble]{.to_tibble()}}.
}
\examples{
letters[1:5] |> rlang::set_names() |> .vector_2_tibble()
letters[1:5] |> rlang::set_names() |> .vector_2_tibble(.rowname = 'a_name', .vec_name = 'alphas')
}
