% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enflatten.R
\name{enflatten}
\alias{enflatten}
\title{Flatten an embedded object list to a single flat list}
\usage{
enflatten(
  ...,
  name_spec = "{outer}|{inner}",
  name_repair = c("minimal", "unique", "check_unique", "universal"),
  .f = ~entibble(.)
)
}
\arguments{
\item{...}{A data object, data objects, or a list of data objects.}

\item{.f}{function to apply at every leaf of the dis-embedded list; default is \code{entibble}}
}
\value{
A flattened list of tibbles.
}
\description{
Recursively flattens an embedded object list
Calls \link{entibble} to coerce internal objects to tibbles
Concatenate embedded-list names with name_spec like '|' or '_'.
}
\examples{
enflatten(iris, mtcars)

list(c(list(1:5), a = list(5:1, 'green', list('blue'))), purrr::set_names(letters, LETTERS)) |> str()
list(c(list(1:5), a = list(5:1, 'green', list('blue'))), purrr::set_names(letters, LETTERS)) |>
  enflatten() |> str()
(embedded_list <- list(list(list(1:5), a = list(5:1, 'green', list('blue')), purrr::set_names(letters, LETTERS)) |> list(head(iris), tail(mtcars))))
embedded_list |> str()
embedded_list |> names()
embedded_list |> enflatten() |> str()
embedded_list |> enflatten() |> names()
embedded_list |> enflatten(name_repair = 'unique') |> names()
embedded_list |> enflatten(name_spec = '{outer}_outer_|{inner}_inner_') |> names()
embedded_list |> enflatten(name_spec = '', name_repair = 'unique') |> names()
list(list(list(1:5), list(5:1, 'green', list('blue')), purrr::set_names(letters, LETTERS)) |>
list(head(iris), tail(mtcars))) |> enflatten(name_spec = 'tbl_{inner}') |> names()
list(list(list(1:5), list(5:1, 'green', list('blue')), purrr::set_names(letters, LETTERS))) |>
list(head(iris), tail(mtcars)) |> enflatten(name_spec = 'lst_grp1.{inner}') |> names()
some_list <- list(list(grp_a = 2:1, grp_b = list(b1 = 2, b2 = list(list(f = iris))), grp_c = list(c2 = 3, c1 = 4)))
enflatten(some_list)

}
