% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_savepath.R
\name{expr_savepath}
\alias{expr_savepath}
\title{Create a savepath from user expression or user specs}
\usage{
expr_savepath(..., .path = NULL, .file_suffix = ".xlsx", .tmp_prefix = "xlr-")
}
\arguments{
\item{...}{an expression (unused if .path is specified)}

\item{.path}{a full save path or file name string with out without suffix}

\item{.file_suffix}{a file type, e.g. '.xlsx' or '.pdf'}

\item{.tmp_prefix}{if producing a temp file, prefix before date and time name components, e.g. 'xlr'}
}
\value{
the complete path string
}
\description{
For use in a function that takes dots
}
\examples{
expr_savepath(.path = 'hi')
iris |> expr_savepath()
iris |> expr_savepath(.path = 'hi')
expr_savepath(iris, .file_suffix = '.pdf', .tmp_prefix = 'boing-')
# .path, if specified, overrides prefix
expr_savepath(.path = 'hi', .file_suffix = '.pdf', .tmp_prefix = 'boing-')
iris |> expr_savepath(.path = 'somewhere/hi/bye', .file_suffix = '.pdf')

}
