% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscript.R, R/format_script.R
\name{format_script}
\alias{format_script}
\title{Adds breaks to the text of a deparsed data structure script}
\usage{
format_script(obj_expr, editor_width = 70)

format_script(obj_expr, editor_width = 70)
}
\arguments{
\item{obj_expr}{text for a data structure}

\item{editor_width}{limit for the number of characters desired in data folding}
}
\value{
a string with embedded '/n' line breaks
}
\description{
Adds breaks to the text of a deparsed data structure script

Adds breaks to the text of a deparsed data structure script
}
