% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_vector.R
\name{get_beta_vector_data}
\alias{get_beta_vector_data}
\title{Get the data from a beta_vector}
\usage{
get_beta_vector_data(x)
}
\arguments{
\item{x}{a beta_vector}
}
\value{
the contents of vector x
}
\description{
Get the data from a beta_vector
}
