% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_vector.R
\name{get_xlr_vector_data}
\alias{get_xlr_vector_data}
\title{Get the data from a xlr_vector}
\usage{
get_xlr_vector_data(x)
}
\arguments{
\item{x}{a xlr_vector}
}
\value{
the contents of vector x
}
\description{
Get the data from a xlr_vector
}
