% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_format.R
\name{is_beta_format}
\alias{is_beta_format}
\title{Test if an object is a \code{beta_format}}
\usage{
is_beta_format(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Test if an object is a \code{beta_format}
}
\examples{
# Test if an object is a beta_format
is_beta_format(1)
bf <- beta_format(font_size = 14)
is_beta_format(bf)
}
