% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_format.R
\name{is_xlr_format}
\alias{is_xlr_format}
\title{Test if an object is a \code{xlr_format}}
\usage{
is_xlr_format(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
a logical.
}
\description{
Test if an object is a \code{xlr_format}
}
\examples{
# Test if an object is a xlr_format
is_xlr_format(1)
bf <- xlr_format(font_size = 14)
is_xlr_format(bf)
}
