% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_savepath.R
\name{locn_reporter}
\alias{locn_reporter}
\title{Report out save location}
\usage{
locn_reporter(path, desc = "")
}
\arguments{
\item{path}{a complete path to cat}

\item{desc}{optional to describe type of path}
}
\value{
the input path
}
\description{
Report out save location
}
\examples{
here::here() |> locn_reporter( desc = 'Workbook ')
}
