% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_integer.R
\name{new_beta_integer}
\alias{new_beta_integer}
\title{Constructor of beta_integer}
\usage{
new_beta_integer(x = integer(), style = beta_format(), call = caller_env())
}
\arguments{
\item{x}{A numeric vector
\itemize{
\item For \code{beta_integer()}: A numeric vector
\item For \code{is_beta_integer()}: An object to test
\item For \code{as_beta_integer()} : a vector
}}

\item{style}{Additional styling options for the vector. See \link{beta_format} for more details.}

\item{call}{the calling environment}
}
\description{
Constructor of beta_integer
}
