% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_percent.R
\name{new_beta_percent}
\alias{new_beta_percent}
\title{Constructor of percent}
\usage{
new_beta_percent(
  x = double(),
  dp = 0L,
  style = beta_format(),
  call = caller_env()
)
}
\arguments{
\item{x}{\itemize{
\item For \code{beta_double()}: A numeric vector
\item For \code{is_beta_double()}: An object to test
\item For \code{as_beta_double()} : a vector
}}

\item{dp}{the number of decimal places to print}

\item{style}{Additional styling options for the vector. See \link{beta_format} for more details.}

\item{call}{the calling environment}
}
\description{
Constructor of percent
}
