% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_vector.R
\name{new_beta_vector}
\alias{new_beta_vector}
\title{Constructor of beta_vector}
\usage{
new_beta_vector(
  x = vector(),
  excel_format = "GENERAL",
  style = beta_format(),
  call = caller_env()
)
}
\arguments{
\item{x}{A vector
\itemize{
\item For \code{beta_vector()}: A vector
\item For \code{is_beta_vector()}: An object to test
\item For \code{as_beta_vector()} : a vector
}}

\item{excel_format}{a character, the excel cell format, not validated. See
\link[openxlsx]{createStyle} argument numFmt for more details on what
you can specify.}

\item{style}{Additional styling options for the vector. See \link{beta_format} for
more details.}

\item{call}{the calling environment}
}
\description{
Constructor of beta_vector
}
