% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_double.R
\name{new_xlr_double}
\alias{new_xlr_double}
\title{Constructor of xlr_double}
\usage{
new_xlr_double(
  x = double(),
  dp = 0L,
  style = xlr_format(),
  call = caller_env()
)
}
\arguments{
\item{x}{\itemize{
\item For \code{xlr_double()}: A numeric vector
\item For \code{is_xlr_double()}: An object to test
\item For \code{as_xlr_double()} : a vector
}}

\item{dp}{the number of decimal places to print}

\item{style}{Additional styling options for the vector. See \link{xlr_format} for more details.}

\item{call}{the calling environment}
}
\description{
Constructor of xlr_double
}
