% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_integer.R
\name{new_xlr_integer}
\alias{new_xlr_integer}
\title{Constructor of xlr_integer}
\usage{
new_xlr_integer(
  x = integer(),
  style = xlr_format_numeric(),
  call = caller_env()
)
}
\arguments{
\item{x}{A numeric vector
\itemize{
\item For \code{xlr_integer()}: A numeric vector
\item For \code{is_xlr_integer()}: An object to test
\item For \code{as_xlr_integer()} : a vector
}}

\item{style}{Additional styling options for the vector. See \link{xlr_format_numeric} for more details.}

\item{call}{the calling environment}
}
\description{
Constructor of xlr_integer
}
