% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_numeric.R
\name{new_xlr_numeric}
\alias{new_xlr_numeric}
\title{Constructor of xlr_numeric}
\usage{
new_xlr_numeric(
  x = double(),
  dp = 0L,
  scientific = FALSE,
  style = xlr_format_numeric(),
  call = caller_env()
)
}
\arguments{
\item{x}{\itemize{
\item For \code{xlr_numeric()}: A numeric vector
\item For \code{is_xlr_numeric()}: An object to test
\item For \code{as_xlr_numeric()} : a vector
}}

\item{dp}{the number of decimal places to print}

\item{scientific}{logical. Whether to format the numeric using scientific notation.}

\item{style}{Additional styling options for the vector. See \link{xlr_format_numeric} for more details.}

\item{call}{the calling environment}
}
\description{
Constructor of xlr_numeric
}
