% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_vector.R
\name{new_xlr_vector}
\alias{new_xlr_vector}
\title{Constructor of xlr_vector}
\usage{
new_xlr_vector(
  x = vector(),
  excel_format = "GENERAL",
  style = xlr_format(),
  call = caller_env()
)
}
\arguments{
\item{x}{A vector
\itemize{
\item For \code{xlr_vector()}: A vector
\item For \code{is_xlr_vector()}: An object to test
\item For \code{as_xlr_vector()} : a vector
}}

\item{excel_format}{a character, the excel cell format, not validated. See
\link[openxlsx]{createStyle} argument numFmt for more details on what
you can specify.}

\item{style}{Additional styling options for the vector. See \link{xlr_format} for
more details.}

\item{call}{the calling environment}
}
\description{
Constructor of xlr_vector
}
