% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_for_xl.R
\name{run_copy_for_xl}
\alias{run_copy_for_xl}
\title{Addin function to call \code{\link[=copy_for_xl]{copy_for_xl()}}}
\usage{
run_copy_for_xl()
}
\description{
For interactive use in RStudio.
}
\examples{
\dontrun{
# Highlight a text item below and press key-chord `ctrl + alt + shift + c`.
# To implement the key-chord in RStudio. Run [set_xlr_key_chords()].
# Note that setting [set_xlr_key_chords()] will restart the R session.

head(iris)
colours()
letters
dplyr::bind_cols(letters, LETTERS)
letters |> rlang::set_names(LETTERS)

# ragged list embedded tibbles should throw an error
enlist(head(iris),letters)

}
}
