% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscript.R
\name{run_enscript}
\alias{run_enscript}
\title{Addin function to call \code{\link[=enscript]{enscript()}}}
\usage{
run_enscript()
}
\value{
No return
}
\description{
Requires RStudio
}
\examples{
\dontrun{
# Highlight each text item below and press key-chord `ctrl + alt + shift + n`.
# To implement quick key-chord. Run {.fn xlr::set_xlr_key_chords}.

(1:5 * 10)
rep("ğŸŠğŸŒˆ",3)
dplyr::starwars |> head()
letters
}

}
