% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_from_xl.R
\name{run_paste_from_xl}
\alias{run_paste_from_xl}
\title{Addin function for user to call \code{\link[=paste_from_xl]{paste_from_xl()}}
Responds to \code{ctrl+alt+shift+v}}
\usage{
run_paste_from_xl()
}
\value{
void
}
\description{
Addin function for user to call \code{\link[=paste_from_xl]{paste_from_xl()}}
Responds to \code{ctrl+alt+shift+v}
}
