% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.R
\name{run_xl}
\alias{run_xl}
\title{background-color: #00A500;
Addin function to call \code{\link[=xl]{xl()}}}
\usage{
run_xl()
}
\value{
void
}
\description{
Requires RStudio
}
\examples{
\dontrun{
# Highlight each item below and press key-chord `ctrl + alt + shift + >`.
# To implement the quick-keys. Run {.fn set_xlr_key_chords}.

mtcars
(1:5 * 10)
rep("ğŸŠğŸŒˆ",3)
dplyr::starwars |> head()
letters
}

}
