% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.R
\name{set_xl_args}
\alias{set_xl_args}
\title{Set persistent default arguments for function \code{xlr::xl()}}
\usage{
set_xl_args(..., reset = FALSE)
}
\arguments{
\item{...}{parameters and arguments from \code{xl()}}

\item{reset}{remove the global \code{xl()} option arguments currently set}
}
\value{
current default args (invisibly)
}
\description{
Reconciles the parameters with the arguments from \code{xl()} and
stores the new defaults session in \code{\link[=options]{options()}}
}
\examples{
(store_options <- getOption('xlr.xl'))

set_xl_args(.quiet = TRUE, .tabname_spec = list(zoom = 65))
getOption('xlr.xl')
set_xl_args(.workbook_spec = list(asTable = FALSE, zoom = 85, withFilter = TRUE))
set_xl_args(reset = TRUE)
getOption('xlr.xl')
set_xl_args(.quiet = FALSE, reset = TRUE)
set_xl_args(.sheet_titles = toupper)

# restore original options
set_xl_args(store_options, reset = TRUE)

}
\seealso{
\code{\link[=xl]{xl()}}
}
