% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_xlr_key_chords.R
\name{set_xlr_key_chords}
\alias{set_xlr_key_chords}
\title{Sets xlr key-chords}
\usage{
set_xlr_key_chords()
}
\description{
\itemize{
\item Sets key-chord "ctrl + alt + shift + v" in RStudio to auto-run \code{\link[=paste_from_xl]{paste_from_xl()}}.
\item Sets key-chord "ctrl + alt + shift + c" in RStudio to auto-run \code{\link[=copy_for_xl]{copy_for_xl()}}.
\item Sets key-chord "ctrl + alt + shift + n" in RStudio to auto-run \code{\link[=enscript]{enscript()}}.
\item Sets key-chord "ctrl + alt + shift + l" in RStudio to auto-run \code{\link[=xl]{xl()}}.
}
}
\details{
To enable the quick keys after running \code{set_xlr_key_chords()}, you must close
and re-launch RStudio.

To set quick-keys manually in RStudio, follow these instructions from Posit:
\url{https://support.posit.co/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}
}
\examples{
\dontrun{ set_xlr_key_chords() }

}
