% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{splitter}
\alias{splitter}
\title{Splits a dataframe into a list of dataframes with auto-naming}
\usage{
splitter(.data, ..., .sep = "|")
}
\arguments{
\item{.data}{a dataframe}

\item{...}{specified columns to split by}

\item{.sep}{character by which to}
}
\value{
named list of dataframes
}
\description{
Similar to \code{dplyr::group_split()} or \code{base::split()}, but:
}
\details{
\itemize{
\item takes only dataframes
\item auto-names the list of output dataframes
\item splits based on column names, column numbers, or any selection features from \code{\link[dplyr:select]{dplyr::select()}}
\item if user specifies no columns to split by, \code{splitter()} will split on groupings from \code{\link[dplyr:group_by]{dplyr::group_by()}}
\item does not create dataframe groups, but preserves dataframe groups in output
}
}
\examples{
iris |> splitter(Species)
iris |> splitter(5)
mtcars |> splitter(mpg)
}
