% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_open.R
\name{sys_open}
\alias{sys_open}
\title{Open a file or folder with the OS default app}
\usage{
sys_open(...)
}
\arguments{
\item{...}{path strings to the files or folder you want to open
-- inputs can be any combination of lists, vectors, or comma-separated inputs
-- no input opens the current working folder location
-- paths are automatically expanded using \code{\link[=normalizePath]{normalizePath()}}}
}
\value{
character vector of named paths
}
\description{
Open files or folders through the operating system default launch application
}
\examples{
\dontrun{
sys_open('.')
sys_open('..')
sys_open('~')
sys_open('/')
sys_open(here::here())
}

}
