% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquify_names.R
\name{uniquify_names}
\alias{uniquify_names}
\title{Glue sequential padded numbers within each group of repeated names}
\usage{
uniquify_names(
  tabnames,
  truncate_side = "right",
  paste_side = "right",
  width = 31,
  sep = "->",
  pad = ".",
  ellipsis = "",
  quiet = FALSE
)
}
\arguments{
\item{tabnames}{character vector of tab/sheet names}

\item{truncate_side}{snip excess tabnames on 'left', 'right', or 'center'}

\item{paste_side}{'left', 'right' for side of the tabnames to tack on rownumbers}

\item{width}{number of chars allowed in the tab/sheet name}

\item{sep}{tabname text and rownumber separator}

\item{pad}{padding characters used on left side of group rownumbering}

\item{ellipsis}{replacement characters during truncation}

\item{quiet}{quiet warning reports}
}
\value{
vector or unique tabnames
}
\description{
Note: tabname cleaning and parameter checks should already have been done in
scrub_tabnames before passing args. This function is recursive and can
call itself forever if width, sep, and pad aren't checked in advance.
}
\examples{
tabnames = c('', NULL, NULL, NA, NA, 'a', 'a', 'a', 'b', 'b', 'c', rep('d',11))
uniquify_names(tabnames)
uniquify_names(tabnames = tabnames, paste_side = 'left' , sep = '|')

}
