% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_vector.R
\name{vec_ptype2.xlr_vector.xlr_vector}
\alias{vec_ptype2.xlr_vector.xlr_vector}
\title{Define generics needed by vctrs to have a well maintained class}
\usage{
\method{vec_ptype2}{xlr_vector.xlr_vector}(x, y, ...)
}
\description{
Define generics needed by vctrs to have a well maintained class
}
