% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.R
\name{xl}
\alias{xl}
\title{View an R objects in a Libreoffice or Excel workbook}
\usage{
xl(
  ...,
  .path = NULL,
  .open = TRUE,
  .quiet = TRUE,
  .sheet_titles = stringr::str_to_title,
  .dataframe_spec = NULL,
  .tabname_spec = list(sep = ".", pad = ".", name_spec = "{inner}"),
  .workbook_spec = list(asTable = TRUE, orientation = "landscape", zoom = 70),
  .return = list(NULL, "savepath", "tibbles", "workbook")
)
}
\arguments{
\item{...}{A dataframe, list of dataframes, or inputs coerceable to one
or more dataframes.}

\item{.path}{Optional path to save a copy of the output spreadsheet. Specify
a minimal file name s*-+
3.6920uch as 'expenses' or a .path such as
'my-data-stash/workbook.xlsx'. \link{.path} uses \code{\link[here:here]{here::here()}}. If .path is not
specified, workbook is removed on or before closure. If .path is specified
but the folder structure does not exist, \code{\link[=xl]{xl()}} will attempt to
construct it.}

\item{.open}{TRUE (default); If FALSE, workbook will not open after being written.}

\item{.quiet}{TRUE or FALSE (default), denoting whether you want messaging.}

\item{.sheet_titles}{NULL, a character vector the same length as the number of
spreadsheets, or a purrr-style function/formula to apply to the default input names.
Default is the function \code{stringr::str_to_title}.}

\item{.dataframe_spec}{NULL (default), or a purrr-style function to apply
across all dataframes; e.g. \verb{~janitor::clean_names(., case = 'title'))} or
\code{janitor::clean_names} to titlecase or snakecase fieldnames.}

\item{.tabname_spec}{a list that allows the user to define tab labeling
arguments: sep = ".", pad = ".", name_spec = "{inner}"}

\item{.workbook_spec}{a list of arguments to pass to
\code{\link[openxlsx:buildWorkbook]{openxlsx::buildWorkbook()}}, e.g. list(asTable = TRUE, orientation =
'landscape', zoom = 70)}

\item{.return}{defines the type of output to R, such as the dataframes, the
workbook-object, or the save path. Default is NULL.}
}
\value{
The side-effect, a workbook of spreadsheets, is the primary output of this function.
The direct-return output is TRUE or FALSE (default) denoting whether the
xlsx workbook file wrote out properly. However, if the parameter \code{.return} is
specified by the user, the function will instead return the list of inputs as
tibbles, the \code{openxlsx} workbook-object, or the save path. All direct outputs
are returned invisibly.
}
\description{
Provides a means of reviewing datasets through the user's default spreasheet
program. Produces output similar to \code{\link[tibble:view]{tibble::view()}}, but in a spreadsheet,
which allows for larger datasets and a list of dataframes.
}
\details{
Auto-corrects forbidden sheetnames
Flattens embedded lists with preserved name structure
}
\examples{
\dontrun{
xl(list('hi'))
enlist('hi') |> xl()
enlist(hi) |> xl()
xl(mtcars)
xl(NA)
xl(a = NA)
xl(character(0))
xl(';')
xl(TRUE)
list(c(list(1:5), list(5:1)), letters) |> xl()
xl(c(1:5))
c(1:5) |> xl()
xl(1)
xl(list(a = 1,5))
list(a = 1,b=5, 1:10, letters, as.matrix(warpbreaks[1:10,]) ) |> xl()
c(1,1:5, list(1:10)) |> xl()
a <- tibble::tibble(iris); xl(a)
rlang::set_names(as.list(letters), LETTERS) |> xl()
a_dataframe <- tibble::tibble(iris); xl(a_dataframe)
xl()
xl('')
xl(NULL)
xl(a = NULL)
xl(iris, dplyr::starwars, mtcars, cars)
}

}
\seealso{
\code{\link[openxlsx:buildWorkbook]{openxlsx::buildWorkbook()}}
}
