% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr-package.R
\docType{package}
\name{xlr-package}
\alias{xlr}
\alias{xlr-package}
\title{xlr: Tools for working interactively between Excel/Libreoffice and R}
\description{
Move data quickly between spreadsheets and R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eauleaf.github.io/xlr/}
}

}
\keyword{internal}
