% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr.R
\name{xlr}
\alias{xlr}
\title{View an R objects in a Libreoffice or Excel workbook}
\usage{
xlr(
  ...,
  .path = NULL,
  .open = TRUE,
  .quiet = TRUE,
  .sheet_titles = TRUE,
  .fieldname_spec = NULL,
  .tabname_spec = list(sep = ".", pad = ".", name_spec = "{inner}"),
  .workbook_spec = list(asTable = TRUE, orientation = "landscape", zoom = 70),
  .return = list(NULL, "savepath", "tibbles", "workbook")
)
}
\arguments{
\item{...}{A dataframe, list of dataframes, or inputs coerceable to one
or more dataframes.}

\item{.path}{Optional path to save a copy of the output spreadsheet. Specify
a minimal file name s*-+
3.6920uch as 'expenses' or a .path such as
'my-data-stash/workbook.xlsx'. \link{.path} uses \code{\link[here:here]{here::here()}}. If .path is not
specified, workbook is removed on or before closure. If .path is specified
but the folder structure does not exist, \code{\link[=xlr]{xlr()}} will attempt to
construct it.}

\item{.open}{If FALSE, workbook will not open after being written.}

\item{.quiet}{TRUE or FALSE (default) denoting whether you want messaging}

\item{.sheet_titles}{TRUE (default), FALSE, or a character vector/list of the
same count as spreadsheets to be written}

\item{.fieldname_spec}{NULL (default) or a purrr-style function to apply
across all fieldnames; e.g. \verb{~janitor::clean_names(., case = 'title'))} to
titlecase fieldnames, 'toupper' for all caps, or \code{janitor::clean_names} for
snakecase}

\item{.tabname_spec}{a list that allows the user to define tab labeling
arguments: sep = ".", pad = ".", name_spec = "{inner}"}

\item{.workbook_spec}{a list of arguments to pass to
\code{\link[openxlsx:buildWorkbook]{openxlsx::buildWorkbook()}}, e.g. list(asTable = TRUE, orientation =
'landscape', zoom = 70)}

\item{.return}{defines the type of output to R, such as the dataframes, the
workbook-object, or the save path. Default is NULL.}
}
\value{
The side-effect, a workbook of spreadsheets, is the primary output of this function.
The direct-return output is TRUE or FALSE (default) denoting whether the
xlsx workbook file wrote out properly. However, if the parameter \code{.return} is
specified by the user, the function will instead return the list of inputs as
tibbles, the \code{openxlsx} workbook-object, or the save path. All direct outputs
are returned invisibly.
}
\description{
TODO: background-color: #00A500;
Provides a means of reviewing datasets through the user's default spreasheet
program. Produces output similar to \code{\link[tibble:view]{tibble::view()}}, but in a spreadsheet,
which allows for larger datasets and a list of dataframes.
}
\details{
Auto-corrects forbidden sheetnames
Flattens embedded lists with preserved name structure
}
\examples{
\dontrun{
xlr(list('hi'))
enlist('hi') |> xlr()
enlist(hi) |> xlr()
xlr(mtcars)
xlr(NA)
xlr(a = NA)
xlr(character(0))
xlr(';')
xlr(TRUE)
list(c(list(1:5), list(5:1)), letters) |> xlr()
xlr(c(1:5))
c(1:5) |> xlr()
xlr(1)
xlr(list(a = 1,5))
list(a = 1,b=5, 1:10, letters, as.matrix(warpbreaks[1:10,]) ) |> xlr()
c(1,1:5, list(1:10)) |> xlr()
a <- tibble::tibble(iris); xlr(a)
rlang::set_names(as.list(letters), LETTERS) |> xlr()
a_dataframe <- tibble::tibble(iris); xlr(a_dataframe)
xlr()
xlr('')
xlr(NULL)
xlr(a = NULL)
xlr(iris, dplyr::starwars, mtcars, cars)
}

}
\seealso{
\code{\link[openxlsx:buildWorkbook]{openxlsx::buildWorkbook()}}
}
