% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_table.R
\name{xlr_and_dplyr}
\alias{xlr_and_dplyr}
\title{xlr and dplyr}
\description{
\code{xlr_table()} is designed to work with dplyr verbs by default. This is so you
\code{mutate}, \code{summarise}, \code{arrange} etc. your data without losing your \code{xlr_table}
information. Particularly if you have used \code{build_table} first on your data,
which outputs data as a \code{xlr_table}.

The list of currently supported dplyrs verbs are: \code{arrange}, \code{distinct}, \code{filter},
\code{mutate}, \code{relocate}, \code{rename}, \code{rename_with}, \code{rowwise}, \code{select}, \code{slice},
\code{slice_head}, \code{slice_max}, \code{slice_min}, \code{slice_sample}, \code{slice_tail}, \code{summarise}.
}
