% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr_numeric.R
\name{xlr_numeric}
\alias{xlr_numeric}
\alias{is_xlr_numeric}
\alias{as_xlr_numeric}
\title{\code{xlr_numeric} vector}
\usage{
xlr_numeric(x = numeric(), dp = 2L, scientific = FALSE, style = xlr_format())

is_xlr_numeric(x)

as_xlr_numeric(x, dp = 0L, style = xlr_format())
}
\arguments{
\item{x}{\itemize{
\item For \code{xlr_numeric()}: A numeric vector
\item For \code{is_xlr_numeric()}: An object to test
\item For \code{as_xlr_numeric()} : a vector
}}

\item{dp}{the number of decimal places to print}

\item{scientific}{logical,}

\item{style}{Additional styling options for the vector. See \link{xlr_format} for more details.}
}
\value{
An S3 vector of class \code{xlr_numeric}
}
\description{
This creates a double vector that represent doubles nicely, so that it is
both printed nicely, and can easily be outputted to excel in the correct format
}
\seealso{
\code{\link[=xlr_percent]{xlr_percent()}}, \code{\link[=xlr_integer]{xlr_integer()}}, \code{\link[=xlr_vector]{xlr_vector()}}
}
