% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_tabnames.R
\name{.check_forbidden_pad}
\alias{.check_forbidden_pad}
\title{Throw error for prohibited padding options or more than 1 char.}
\usage{
.check_forbidden_pad(pad = "")
}
\arguments{
\item{arg}{'pad' in function scrub_tabnames()}
}
\value{
a single character; throws an error if input is any of these characters \/:?*'[]
}
\description{
Throw error for prohibited padding options or more than 1 char.
}
\examples{
.check_forbidden_pad('.')
.check_forbidden_pad('-')
\dontrun{ # expect error if any of these: \/:?*'[]
.check_forbidden_pad(':')
.check_forbidden_pad('')
.check_forbidden_pad(NA)
.check_forbidden_pad(NULL)
.check_forbidden_pad("\\\\")
}

}
\seealso{
\code{\link[=scrub_tabnames]{scrub_tabnames()}}
}
