% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_tabnames.R
\name{.fix_forbidden_tabnames}
\alias{.fix_forbidden_tabnames}
\title{Corrects tabnames by replacing special characters one-for-one
and notifies the user of replaced characters.}
\usage{
.fix_forbidden_tabnames(tabnames, quiet = FALSE)
}
\arguments{
\item{tabnames}{vector of tab/sheet names}

\item{quiet}{TRUE or FALSE to suppress notices of forbidden-char replacement}
}
\value{
the input vector with forbidden characters replaced/removed
}
\description{
Quietly casts input as characters.
}
\note{
NA always returned as NA
NULL acts like c(), i.e c(NULL, NULL)
}
\examples{
c("' '?HisTory,*?*!@#$\%^&*()_+ '[{1'}]'[:/\\\\]''", "", NA, NULL) |>  .fix_forbidden_tabnames()

" ''''''" |> .fix_forbidden_tabnames()
"'''1' '" |> .fix_forbidden_tabnames()
"'[ ]hi];'" |> .fix_forbidden_tabnames()
"!@#$\%^&*()_+ {HI} ^" |> .fix_forbidden_tabnames()
"'histor'y'" |> .fix_forbidden_tabnames()
"[:/] " |> .fix_forbidden_tabnames()
"?/a\\\\" |> .fix_forbidden_tabnames()
"?,*?*?'" |> .fix_forbidden_tabnames()
"[history coursework]" |> .fix_forbidden_tabnames()
"'HISTORIC' History '''" |> .fix_forbidden_tabnames()
"*[history' buff]*' " |> .fix_forbidden_tabnames()
NA |> .fix_forbidden_tabnames()
NULL |> .fix_forbidden_tabnames()
c(NULL, NA, '[hi?]') |> .fix_forbidden_tabnames()
c(NULL, NULL) |> .fix_forbidden_tabnames()
enlist(NULL, NA, '[hi?]') |>  purrr::map(.fix_forbidden_tabnames)

}
\seealso{
\code{\link[=scrub_tabnames]{scrub_tabnames()}}

\code{\link[=.forbidden_chars_replace]{.forbidden_chars_replace()}}
}
