% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscript.R
\name{.format_script}
\alias{.format_script}
\title{Adds breaks to the text of a deparsed data structure script}
\usage{
.format_script(obj_expr, snip_width = 70)
}
\arguments{
\item{obj_expr}{text for a data structure}

\item{snip_width}{limit for the number of characters desired before data folding at comma-space breaks}
}
\value{
a string with embedded '/n' line breaks
}
\description{
Adds breaks to the text of a deparsed data structure script
}
\examples{
letters[1:5] |> deparse1(backtick = TRUE) |> xlr:::.format_script()
list(tail(iris,2)) |> deparse1(backtick = TRUE) |> xlr:::.format_script()

}
