% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entibble.R
\name{entibble}
\alias{entibble}
\title{Convert to tibble with rownames}
\usage{
entibble(
  ...,
  .rowname = "rowname",
  .name_repair = c("minimal", "check_unique", "unique", "universal", "unique_quiet",
    "universal_quiet")
)
}
\arguments{
\item{...}{object or expression to convert to a tibble}

\item{.rowname}{String to name the column of rownames if passed a dataframe or matrix with rownames, or a vector with names; default is column name is 'rowname'. If .rownames = NULL, \code{\link[=entibble]{entibble()}} acts identical to \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{.name_repair}{}
}
\value{
a tibble with rownames in a column
}
\description{
Same as \code{\link[tibble:tibble]{tibble::tibble()}}, except it automatically includes a rownames
column if passed a named dataframe, matrix, or vector.
}
\examples{
letters |> entibble()
letters |>  purrr::set_names() |> entibble()
letters |>  purrr::set_names() |> entibble(.rowname = 'something from user')
letters |>  purrr::set_names() |> entibble(.rowname = 'something from user', .name_repair = ~c('hello', 'goodbye'))
1:5 |> entibble()
tibble::tibble(y = 1:3,x = 11:13)
entibble(y = c(1, 2, 3),x = c(11, 12, 13))
tibble::tibble(x = 1:3,x = 11:13)
entibble(x = 1:3, x = 11:13)
tibble::tibble(x = 1:3,x = 11:13, .name_repair = 'minimal')
entibble(x = 1:3,x = 11:14, .name_repair = 'minimal')
tibble::tibble(x = 1:3,x = 11:14, .name_repair = 'minimal')
tibble::tibble(x = 1:3,x = 11:13, .name_repair = 'unique')
entibble(x = 1:3,x = 11:13, .name_repair = 'unique')
entibble(x = 1:3,x = 11:13, .name_repair = 'unique_quiet')
entibble(x = 1:3,x = 11:13, .name_repair = ~scrub_tabnames(., sep = ' <-- ', quiet = T))
tibble::tibble(x = 1:3,x = 11:13, .name_repair = ~scrub_tabnames(., sep = ' <-- ', quiet = T))

as.matrix(1:10) |> rownames()
as.matrix(1:10) |> entibble()
as.matrix(1:10) |> tibble::tibble()
as.matrix(warpbreaks[1:10, ]) |> rownames()
as.matrix(warpbreaks[1:10, ]) |> entibble()
as.matrix(warpbreaks[1:10, ]) |> entibble(.rowname = 'row_nums')
matrix(c(1, 2, 3, 11, 12, 13), nrow = 2, ncol = 3, byrow = TRUE, dimnames = list(c('row1', 'row2'), c('C.1', 'C.2', 'C.3'))) |> entibble()
tail(iris) |> tibble::has_rownames()
tail(iris) |> entibble()
mtcars |> entibble()
mtcars |> entibble(.rowname = 'make')
5 |> entibble()
NA |> entibble()
NULL |> entibble()
enlist(head(iris), head(mtcars)) |> purrr::map(entibble)
enlist(head(iris), head(mtcars)) |> entibble()
head(iris) |> purrr::map(entibble)
entibble(!!!head(iris))
enlist(head(iris), head(mtcars)) |> entibble()
enlist(head(iris), head(mtcars)) |> tibble::tibble()

letters |> purrr::set_names(LETTERS) |> tibble::tibble()
letters |> purrr::set_names(LETTERS) |> entibble()

letters |> purrr::set_names(LETTERS) |> as.list() |> tibble::tibble()
letters |> purrr::set_names(LETTERS) |> as.list() |> entibble()
letters |> purrr::set_names(LETTERS) |> tibble::enframe( name = 'row', value = 'blue')

# If rowname is specified, but the dataframe has no rownames, rowname is ignored.
entibble(iris, , .rowname = 'flower_name')
letters |> as.list() |> tibble::tibble()
letters |> as.list() |> entibble(.rowname = 'a-z')
letters |> list() |> tibble::tibble()
letters |> list() |> entibble(.rowname = 'a-z')
letters |> enlist() |> entibble(.rowname = 'a-z')

colours() |> rlang::set_names() |> entibble() |> entibble(another_col = colours())
colours() |> rlang::set_names() |> entibble(another_col = colours())
letters |> as.list() |> enlist() |> entibble()
dplyr::bind_cols(letters, LETTERS) |> copy_for_xl()
letters |> rlang::set_names(LETTERS) |> copy_for_xl()

entibble(1:10,11:20)
entibble(c(1:10,11:20))
entibble(list(1:10,11:20))
entibble(enlist(1:10,11:20))
entibble(enlist(1:10,11:20), .rowname = 'nums')
entibble(list(1:10),list(11:20))
tibble::tibble(enlist(1:10),enlist(11:20))
entibble(enlist(1:10),enlist(11:20))
entibble(enlist(1:10),enlist(11:20), .rowname = 'blue')
entibble(enlist(1:10),enlist(11:20), .rowname = 'a column name', .name_repair = 'unique')
entibble(enlist(1:10),enlist(11:20), .rowname = 'a column name', .name_repair = 'universal')
entibble(enlist(1:10),enlist(11:20), .rowname = 'a column name', .name_repair = 'universal_quiet')
entibble(enlist(list(1:10),list(11:20)))
entibble(enlist((1:10),(11:20)))

# test
as.matrix(warpbreaks[1:10, ]) |> asplit(2) |> entibble()

}
\seealso{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
