% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_from_xl.R
\name{paste_from_xl}
\alias{paste_from_xl}
\title{Paste formatted data into Rstudio from a spreadsheet, with key chord
'ctrl+alt+shift+v'.}
\usage{
paste_from_xl(has_fieldnames = NULL)
}
\arguments{
\item{has_fieldnames}{TRUE or FALSE indicating whether field names are
present in the spreadsheet data you copied. If user supplies no value,
\code{\link[=paste_from_xl]{paste_from_xl()}} guesses by looking at the clipboard data.

Because parameters are not user settable when employing quick keys, go to the
console window and write 'T' or 'F' in the function echoed to the console
if the function incorrectly guessed the presence of field names.}
}
\value{
-- If user copied a single spreadsheet row into the clipboard memory,
returns a vector.
** If user copied more than one spreadsheet row to the clipboard, returns a tibble.
** If user copied a local file to the clipboard, returns a string with the path to the file.
}
\description{
Pastes spreadsheet data from the user's clipboard to RStudio,
formatted in 1 of 2 ways.
\enumerate{
\item If pasting to the console window, 'ctrl+alt+shift+v' will insert a tibble
of the copied spreadsheet data.
\item If pasting to the text editor window, 'ctrl+alt+shift+v' will insert an
expression to produce the tibble.
}

Will also paste folder and file paths if user pressed 'ctrl + c' on the file
icons in a browser window.

To set the quick keys.
}
\details{
To set the quick keys in RStudio, run function:
\code{\link[=set_xlr_key_chords]{set_xlr_key_chords()}}.
Or, or set the quick keys manually by following these instructions:
\url{https://support.posit.co/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}
}
\note{
Clipboard memory is limited; if you're copying a lot of data from a
spreadsheet, try reading it in instead with \code{\link[readxl:read_excel]{readxl::read_excel()}}.
}
\examples{
\dontrun{
# (All examples require the user to interactively select data.)
# Copy data from a spreadsheet into working memory with 'ctrl + c' or by
# right-clicking with the mouse and selecting 'copy', then type into the R
# console:

(my_data <- paste_from_xl())
(my_data <- paste_from_xl(T))
(my_data <- paste_from_xl(F))
paste_from_xl()

To paste the copied spreadsheet data into R.
}

}
\seealso{
\code{\link[=set_xlr_key_chords]{set_xlr_key_chords()}}
}
