% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_from_xl.R
\name{run_paste_from_xl}
\alias{run_paste_from_xl}
\title{Rstudio addin function to call \code{\link[=paste_from_xl]{paste_from_xl()}}}
\usage{
run_paste_from_xl()
}
\value{
void
}
\description{
For key-chord \code{ctrl+alt+shift+v}.
}
\examples{
\dontrun{ run_paste_from_xl() }

}
