% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_xlr_key_chords.R
\name{set_xlr_key_chords}
\alias{set_xlr_key_chords}
\title{Sets key chord "ctrl + alt + shift + v" to auto-run \code{\link[=paste_from_xl]{paste_from_xl()}}.
Sets key chord "ctrl + alt + shift + c" to auto-run \code{\link[=copy_from_xl]{copy_from_xl()}}.
Restarts rstudio session.}
\usage{
set_xlr_key_chords()
}
\description{
Running this function allows the user to press the key chord "ctrl + alt + shift
\itemize{
\item v" to paste data copied from a spreadsheet into R directly where the cursor sits,
as well as allows the user to press "ctrl + alt + shift + c" to copy a data objects
into memory by variable or expression hightlighted in rstudio.
}
}
\details{
To set quick-keys manually, follow these instructions from Posit:
\link{https://support.posit.co/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}
}
\examples{
\dontrun{ xlr::set_xlr_key_chords() }

}
\seealso{
\code{\link[=paste_from_xl]{paste_from_xl()}}

\code{\link[=copy_for_xl]{copy_for_xl()}}
}
