% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.R
\name{xl}
\alias{xl}
\title{View an R dataframe or dataframes quickly in a Libreoffice or Excel workbook.
TODO: background-color: #00A500;}
\usage{
xl(
  ...,
  .path = NULL,
  .open = TRUE,
  .quiet = FALSE,
  .field_spec = NULL,
  .tabname_spec = list(sep = ".", pad = "."),
  .flatten_spec = "{outer}|{inner}",
  .workbook_spec = list(asTable = TRUE, orientation = "landscape", zoom = 65, start_row =
    3),
  .return = c("tibble-list", "workbook-object")
)
}
\arguments{
\item{...}{A dataframe, list of dataframes, or inputs coerceable to one
or more dataframes.}

\item{.path}{Optional path to save a copy of the output spreadsheet. Specify
a minimal file name such as 'expenses' or a .path such as
'my-data-stash/workbook.xlsx'. \link{.path} uses \code{\link[here:here]{here::here()}}. If .path is not
specified, workbook is removed on or before closure. If .path is specified
but the folder structure does not exist, \code{\link[=xl]{xl()}} will attempt to
construct it.}

\item{.open}{If FALSE, workbook will not open after being written.}

\item{.field_spec}{function #titlecase_fields}

\item{.return}{Type of output in R. Default is the tibbled dataframes. If workbook-object, then}
}
\value{
A force-named flat list of tibbles, returned invisibly, and a workbook of spreadsheets as side-effect
}
\description{
Provides a means of reviewing datasets through the user's default spreasheet
program with opinionated formatting. Produces output similar to
\code{\link[tibble:view]{tibble::view()}}, but in a spreadsheet, allowing for larger datasets and a
list of dataframes. \code{\link[=xl]{xl()}} auto-corrects issues like forbidden sheetname
characters to produce an immediately viewable dataset.
}
\examples{
\dontrun{
xl(mtcars)
xl(a = NULL)
xl(a = NA)
xl(NULL)
list(c(list(1:5), list(5:1)), letters) |> xl()
c(1:5) |> xl()
xl(c(1:5))
xl(1)
xl(list(a = 1,5))
list(a = 1,b=5, 1:10, letters, as.matrix(warpbreaks[1:10,]) ) |> xl()
c(1,1:5, list(1:10)) |> xl()
a <- tibble::tibble(iris); xl(a)
xl(NA)
xl(NULL)
xl(1)
1:26 |> purrr::set_names(letters) |> xl()
as.matrix(warpbreaks[1:10,])
}
specification args

}
\seealso{
\code{\link[openxlsx:buildWorkbook]{openxlsx::buildWorkbook()}}
}
