% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr-package.R
\docType{package}
\name{xlr-package}
\alias{xlr}
\alias{xlr-package}
\title{xlr: Tools for working interactively between Excel and R}
\description{
'xlr' is a group of convenience functions for data scientists who use both Excel and R. The goal is to make it trivial to interactively move data back and forth.
}
\keyword{internal}
