% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr-package.R
\docType{package}
\name{xlr-package}
\alias{xlr}
\alias{xlr-package}
\title{xlr: Tools for working interactively between Excel and R}
\description{
'xlr' is a group of functions for data scientists who use both spreadsheets and R. 'xlr' makes moving data back and forth between spreadsheets and R easier.
}
\keyword{internal}
