% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr-package.R
\docType{package}
\name{xlr-package}
\alias{xlr}
\alias{xlr-package}
\title{xlr: Tools for working between Excel/Libreoffice and R}
\description{
Move data quickly between spreadsheets and R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eauleaf.github.io/xlr/}
  \item \url{https://github.com/eauleaf/xlr}
  \item Report bugs at \url{https://github.com/eauleaf/xlr/issues}
}

}
\author{
\strong{Maintainer}:  Eauleaf \email{eauleaf@gmail.com}

}
\keyword{internal}
